Updated with newest Polymer source.
I took the newest source of ´dom-repeat` and modified it. Added the `_isStarted` property at some splaces to prevent templatizer from running as long as no template was set. Renamed the the `ready`function to `setTemplate`and set `_isStarted=true`. That nearly all. The rest is small modifications to use the code as a behavior, not an element.